Use constant padded entity value size
On further testing, using nearest larger power of
2 sized entries didn't help the fragmentation
problem at all. It turns out to be an issue with
the nginx slab allocator when allocating large
amounts of space. This workaround appears to be the
only tenable approach until this patch
http://forum.nginx.org/read.php?29,240420,241321#msg-241321
or some version of it makes it into nginx mainline.